+++ res/shaders/shaderVertex.glsl
@@ -21,12 +21,17 @@ uniform float useFakeLighting;
 uniform float numberOfRows;
 uniform vec2 offset;
 
+uniform vec4 clipPlane;
+
 const float density = 0.0035;
 const float gradient = 2.6;
 
 void main(void) {
     
     vec4 worldPosition = transformationMatrix * vec4(position, 1.0);
+
+    gl_ClipDistance[0] = dot(worldPosition, clipPlane);
+
     vec4 positionRelativeToCamera = viewMatrix * worldPosition; 
     gl_Position = projectionMatrix * positionRelativeToCamera;
     pass_textureCoords = (textureCoords/numberOfRows) + offset;