+++ res/shaders/terrainVertex.glsl
@@ -16,12 +16,17 @@ uniform mat4 viewMatrix;
 uniform mat4 viewMatrixInv;
 uniform vec3 lightPosition[4];
 
+uniform vec4 clipPlane;
+
 const float density = 0.0035;
 const float gradient = 2.6;
 
 void main(void) {
     
     vec4 worldPosition = transformationMatrix * vec4(position, 1.0);
+
+    gl_ClipDistance[0] = dot(worldPosition, clipPlane);
+
     vec4 positionRelativeToCamera = viewMatrix * worldPosition; 
     gl_Position = projectionMatrix * positionRelativeToCamera;
     pass_textureCoords = textureCoords;